requirements.txt
django==2.1.4
requests==2.20.1
pycryptodome==3.4.7